removed nonsense contrib, moved default tomes
