--- implants/contrib/example.tome
@@ -1,1 +1,0 @@
-file.append("/tmp/do", "stuff")
