+++ tavern/tomes/embed.go
