+++ tavern/tomes/parse.go
