+++ tavern/tomes/parse_test.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"testing"
 
-	"github.com/kcarretto/realm/contrib/tomes"
 	"github.com/kcarretto/realm/tavern/ent/enttest"
 	"github.com/kcarretto/realm/tavern/ent/tome"
+	"github.com/kcarretto/realm/tavern/tomes"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
